+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 - `failure` dependency.
 - Unused error types.
+- `NodeType::Declaration`, `Node::is_declaration`. 
 - `parse_comments`, `parse_declarations`, `parse_unknown_elements`, `parse_unknown_attributes`,
   `parse_px_unit`, `skip_elements_crosslink`, `skip_paint_fallback` from `ParseOptions`.
 - `assert_eq_text` macro.